Provide ability to retrieve token from every remote unit
When the remote charm has multiple units the relation may
have transient state whereby a mix of old and new tokens
are available i.e. eventually consistent. Since tokens are
one-shot and we cant tell good from bad, we need to have
all of them to be able to find a good one.

Related-Bug: #1849323