Publish vault_url when publish_url is called
Previous attempt to solve the external/internal binding problem made it
possible to call the interface's publish_url but not actually set
vault_url if subnets happen to not match. This can be the case in a
number of instances including CMR, AWS and multiple spaces.

This change simplifies the interface and guarantees when publish_url is
called that vault_url is set.

The address selection logic should reside in the vault charm.

Closes-Bug: #1895185